Split manual self-repair precondition tests
Move the manual precondition repair tests out of the end-to-end repair test
module so `repair_end_to_end.rs` stays below the repository file-size limit.
Register the new sibling test module explicitly to match the existing test
module layout.
